requirements/base.txt
pytz==2023.3.post1  # https://github.com/stub42/pytz
Pillow==10.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise
redis==5.0.1  # https://github.com/redis/redis-py
hiredis==2.2.3  # https://github.com/redis/hiredis-py
# celery==5.2.7  # pyup: < 6.0  # https://github.com/celery/celery
celery==5.3.6  # pyup: < 6.0  # https://github.com/celery/celery
flower==2.0.1  # https://github.com/mher/flower
xmltodict==0.13.0  # https://github.com/martinblech/xmltodict.git

# Django
# ------------------------------------------------------------------------------
# django==4.1.8  # pyup: < 4.0  # https://www.djangoproject.com/
django==5.0.3
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.4.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.61.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.1  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2024.2 # https://github.com/django-crispy-forms/crispy-bootstrap5
django-compressor==4.4  # https://github.com/django-compressor/django-compressor
django-redis==5.4.0  # https://github.com/jazzband/django-redis4

# Django REST
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1  # https://django-rest-framework-simplejwt.readthedocs.io/en/latest/

# Wagtail
# ------------------------------------------------------------------------------
wagtail==5.2.3  # https://github.com/wagtail/wagtail


# Wagtail Recaptcha
# ------------------------------------------------------------------------------
# django-recaptcha==4.0.0
wagtail-django-recaptcha==2.1.1

# Wagtail Menu
# ------------------------------------------------------------------------------
wagtailmenus==3.1.9

# Wagtail Localize
# ------------------------------------------------------------------------------
wagtail-localize==1.8.2

#Wagtail Admin
# ------------------------------------------------------------------------------
wagtail-modeladmin==1.0.0

# Django celery
# ------------------------------------------------------------------------------
django-celery-beat==2.6.0  # https://github.com/celery/django-celery-beat
django_celery_results==2.5.1

# Wagtail-Autocomplete
# https://github.com/wagtail/wagtail-autocomplete
# ------------------------------------------------------------------------------
wagtail-autocomplete==0.11.0


# Minio
minio==7.2.5

# packtools
# ------------------------------------------------------------------------------
lxml==4.9.4 # https://github.com/lxml/lxml
tornado>=6.3.3 # not directly required, pinned by Snyk to avoid a vulnerability
-e git+https://github.com/scieloorg/packtools.git@4.11.20#egg=packtools
# packtools@https://github.com/scieloorg/packtools/archive/refs/tags/3.3.3.zip
#-e git+https://github.com/scieloorg/packtools#egg=packtools


# Sickle
# ------------------------------------------------------------------------------
Sickle==0.7.0

# Solr
# ------------------------------------------------------------------------------
django-haystack==3.2.1

# PySolr
# ------------------------------------------------------------------------------
pysolr==3.9.0

# Tenacity
# ------------------------------------------------------------------------------
tenacity==8.2.3  # https://pypi.org/project/tenacity/

#Django Maintenance Mode
# ------------------------------------------------------------------------------
django-maintenance-mode==0.21.1

# Snky
# ------------------------------------------------------------------------------
certifi>=2023.7.22 # not directly required, pinned by Snyk to avoid a vulnerability
requests>=2.31.0 # not directly required, pinned by Snyk to avoid a vulnerability

python-docx==1.1.2
huggingface-hub==0.26.1
llama-cpp-python==0.3.1
google-generativeai==0.3.2
langid